Seperate friendlist page into components.
Seperated the friendlist page into components. The page component is now
called FriendListPage.  The FriendListPage component is a container
component that contains the FriendList component.  The FriendList
component is a presentational component that contains FriendListItem
components.  The FriendListItem component is a presentational component
that displays a single friend.
