+++ src/app/globals.css
@@ -41,7 +41,7 @@
     --background: 33 10% 17%;
     --foreground: 0 0% 98%;
 
-    --card: 240 10% 3.9%;
+    --card: #f8f5f1;
     --card-foreground: 0 0% 98%;
 
     --popover: 240 10% 3.9%;
